Casting the average calculation to double

+++ SampleCSharp7/FirstConsoleApp/Student.cs
@@ -7,8 +7,8 @@ namespace FirstConsoleApp
         public (double average, int studentCount, bool belowAverage) 
             GetAverageAndCount(int[] scores, int threshold)
         {
-            var result = (average:0.0, studentCount:0, belowAverage:true);
-            result = ((scores.Sum()/scores.Length), scores.Length, result.average.CheckIfBelowAverage(threshold));
+            var result = (average:0D, studentCount:0, belowAverage:true);
+            result = (((double)scores.Sum()/scores.Length), scores.Length, result.average.CheckIfBelowAverage(threshold));
 
             return result;
         }
